src/integracao/requirements.txt
aiohttp==3.9.1
websockets==12.0
redis==5.0.1
msgpack==1.0.7
pyyaml==6.0.1
python-jose==3.3.0
cryptography==41.0.7
prometheus-client==0.19.0
python-json-logger==2.0.7
tenacity==8.2.3
aiokafka==0.8.1
uvicorn==0.24.0
fastapi==0.104.1
pydantic==2.5.2
python-multipart==0.0.6
email-validator==2.1.0.post1
python-dotenv==1.0.0 